requirements.txt
absl-py==0.7.1
argcomplete==1.9.5
asn1crypto==0.24.0
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
Brlapi==0.6.6
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
checkbox-support==0.39.0
cloudpickle==0.8.1
cryptography==2.6.1
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.6
dask==1.1.5
decorator==4.4.0
defer==1.0.6
defusedxml==0.5.0
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
feedparser==5.2.1
gast==0.2.2
Glances==3.1.0
grpcio==1.19.0
h5py==2.9.0
html5lib==1.0.1
httplib2==0.9.2
idna==2.8
imageio==2.5.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
jeepney==0.4
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==19.0.1
keyrings.alt==3.1.1
kiwisolver==1.0.1
language-selector==0.1
louis==3.5.0
lxml==4.3.3
macaroonbakery==1.2.1
Mako==1.0.8
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mock==2.0.0
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.8
numpy==1.16.2
nvidia-ml-py3==7.352.0
oauthlib==3.0.1
opencv-python==4.0.1.23
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plainbox==0.38.0
progressbar==2.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==5.6.1
ptyprocess==0.6.0
pyasn1==0.4.5
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
pygobject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.0
pyRFC3339==1.1
pyrsistent==0.14.11
python-apt==1.6.3+ubuntu1
python-dateutil==2.8.0
python-debian==0.1.34
pytz==2018.9
PyWavelets==1.0.2
pyxdg==0.25
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
requests-unixsocket==0.1.5
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
scour==0.37
SecretStorage==3.1.1
Send2Trash==1.5.0
SimpleITK==1.2.0
six==1.12.0
sklearn==0.0
skorch==0.5.0.post0
soupsieve==1.9
system-service==0.3
systemd-python==234
tabulate==0.8.3
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toolz==0.9.0
torch==1.0.1.post2
torchvision==0.2.2.post3
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
urllib3==1.24.1
virtualenv==16.4.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
xkit==0.0.0
